HTML - PHP Form To Input Student Info
This is a HTML form that allows the user to input student information.
When the user submits the form, the PHP will check the form fields for
errors.  For instance,   First Name and Last Name uses POSIX REGEX to
remove all spaces, punctuation, and digits.  Then, after the removal, an
if statement checks to see if the line is empty.  If the line is empty,
obviously there were no alphabetical characters input by the user.  If
the First and Last Name  are valid, then it uses a Regex pattern to
validate the Student ID and GPA.  If all input the the form fields are
valid, then the information is appended to a text file in the current
working directory as so  "LastName:FirstName:UTCID:GPA".  After the
information is appended, the file is then sorted alphabetically by Last
Name.  That information is then presented in a nicely formatted HTML
table.