PHP-HTML Array Demo
This demonstrates the use of arrays in PHP.  The HTML file contains a
form that has a submit and reset button as well as 3 input form fields.
The form asks for an array size, minimum random value, and maximum
random value.  The PHP then processes the data in the array using a
foreach loop.  For each element in the array it calculates the
following:
-The square root of the absolute value of the element.
-The element squared.
-The element raised to a random exponent.
-The string "positive" if it is positive, "negative" if it is negative,
and "zero" if  0.
-The area of a circle using the absolute value of the element in the
array as the radius.
-The volume of a sphere.